+++ lib/generators/happy_seed/admin/admin_generator.rb
@@ -10,7 +10,7 @@ module HappySeed
         gem 'dateslices'
 
         Bundler.with_clean_env do
-          run "bundle install > /dev/null"
+          run "bundle install --without production"
         end
 
         generate 'active_admin:install'