+++ lib/generators/happy_seed/devise_invitable/devise_invitable_generator.rb
@@ -17,7 +17,7 @@ module HappySeed
         gem 'devise_invitable'
 
         Bundler.with_clean_env do
-          run "bundle install > /dev/null"
+          run "bundle install --without production"
           run 'rake db:migrate'
         end