+++ lib/generators/happy_seed/instagram/instagram_generator.rb
@@ -12,7 +12,7 @@ module HappySeed
         gem 'instagram'
 
         Bundler.with_clean_env do
-          run "bundle install > /dev/null"
+          run "bundle install --without production"
         end
 
         directory "docs"